Fix quoting issues with Windows cleanup
vim/markedly.vim: Passing the filename to the cleanup command in the
autocmd! command is fraught with quoting issues: the function call wants
fnameescape but the subsequent shell command wants shellescape. Solve
the problem by storing the filename in a dictionary keyed by the tag,
then the cleanup can pull out the unquoted file name and shellescape it
happily.