Update to reflect Vundle and vim-airline plugin remotes moving on github
vimrc: Vundle has moved to VundleVim/Vundle.vim, and vim-airline has moved to vim-airline/vim-airline and themes are now separate in vim-airline-themes.

Set g:ruby_path to prevent the ruby ftplugin from messing with &path.

Check for the presence of wildignore before trying to set it in order to cope with older vim versions.