Switch from using the ruby markedly gem to markdown-preview.vim
vim/markedly.vim: Reimplement completely so that if the vim-marked
plugin is available that is used, otherwise the markdown-preview.vim
plugin is used if that is installed. In both cases MarkdownPreview is
added as the command to open the preview, and in both cases it will
auto-close if the buffer closes.

vim/markedly_stop.ps1: Deleted. This helper is no longer needed.

vimrc: Add iamcco/markdown-preview.vim.