Improve the reliability of install-mac.sh
install-mac.sh: An enduring problem with using “brew linkapps” followed
by replacing the symlinks into /usr/local/Cellar with aliases (so
Spotlight indexes them) is that Finder seems to regularly misname the
alias with a suffix of “ alias”. In any case, why ask brew to create
symlinks when they are going to be replaced? This reworks the code to
directly find the .app files in /usr/local/Cellar and create aliases
for them in /Applications. Also improve the reporting.