gvimrc: Font sizing in MacVim, or perhaps macOS in general, works out a bit differently to on Windows and Linux.
vim/plugins.vim: Disable mucomplete's auto completion which doesn't work right in MacVim (GUI), and turns out to be generally annoying elsewhere. It's easy to toggle on with MUcompleteAutoToggle when I want it.

install-mac.sh: Fiddle with the alias renaming code. I should really rewrite this in a proper language.