environment: Do not export all path-like variables
I did not explicitly intend to export CDPATH, but by using add_to_path_var that's what was happening.  Unfortunately a curiously written Makefile in one of our projects broke because of this, ending up trying to execute the stdout from "cd blah".  The only variables that I need to export are PYTHONPATH and RUBYLIB when they are present.