tmux: Updates to configuration after more time using multiple client OSs.
tmux.conf: I had misunderstood tmux-sensible's preference of emacs keys: it is only doing that for the <prefix>-: command line which is fine as tmux has limited vi keys support on the command line.  Detect when running under iTerm and only set-clipboard on in that case as gnome-terminal and PuTTY don't support OSC 52 escape sequences to set/access the clipboard contents.  Add tmux-yank and configure for vi keys.

mac/Brewfile: Add reattach-to-user-namespace for tmux-yank on macOS.

functions.bash: Add "tmux conf" which will split the current window based on the selected session layout, similar to "tmux go" but without creating a new session.  Add "tmux update-environment" to apply the current session variables to the running shell.  Refactor set_title() with a separate __sc_prompt_command() function, and use that to update the environment (as well as the window title).  This means that switching clients and ensuring e.g. $DISPLAY is updated is as simple as attaching and hitting return.