bash-completion: Add bash-completion which Apple seem to ignore.
mac/Brewfile: Add to the mix.

functions.bash: Conditionally load when on Darwin and bash_completion is installed.  Homebrew suggests directly sourcing the main bash_completion script, but I have chosen to use the profile.d wrapper as that includes checks for the correct bash version, that this is an interactive shell, and prevents recursive loading.