functions.bash: Improve window title when under tmux or ssh'd to a server
Why am I trying to mess about falling back to the VTE terminal functions which come with gnome-terminal, and the adding special cases other environments when I could simply set the window title I really want in the first place?  This borrows the guts of __vte_prompt_command() from vte.sh which ships with gnome-terminal, but I have removed the use of OSC 7 which is specific to VTE based terminals.

See the following for more information on gnome-terminal's use of OSC 7:
https://bugs.launchpad.net/ubuntu/+source/gnome-terminal/+bug/1132700
http://midnight-commander.org/ticket/3088