environment: Fix add_to_path_var() for unexported variables
My naive previous change doesn't work because if the variable isn't exported it won't get included in the environment passed to ruby.  Solve that by explicitly passing the current value of the variable, remembering that this may run with POSIX shells like dash, and so cannot use the ${!var} bashism.