Replace airline with a locally defined statusline
vim/plugins.vim: Remove airline and it's themes, and remove the configuration.

vimrc: Rename WrapDescribeForStatusLine().

vim/statusline.vim: The new statusline code.  The status line is set per window with setlocal in order to distinguish active and inactive windows.  When changing this remember the different environment that %!function() and %{function()} are run in: only the latter is run with the environment of the target window.