statusline: use colour to show modified windows
vim/statusline.vim: Dynamically colour section C to indicate modified buffers.  This is done by creating buffer specific highlight groups which link to the underlying statusline_c, statusline_c_mod and statusline_c_mod_inactive groups. The buffer groups are created out of a function %{} called from the status line, and are cleared by an autocmd on BufDelete and BufUnload.

To prevent lots of expensive :highlight! commands when updating the status line, the current &modified state is cache in a buffer-local variable, so the highlight group is only (re)defined when the modified state changes.

Highlight group changes appear to be asynchronous to the statusline update, so this maintains both active and inactive highlight groups for each buffer which avoids the race when changing (in)active state.