vim/statusline.vim: Simplification of the modified colouring.
Replaced the previous highlight group gymnastics with the simple idiom from https://github.com/vim/vim/issues/1697#issuecomment-380216189 which allows for one set of highlight groups with different windows selecting which to use based on %{...} expressions.  The only downside to this approach is that the "active window with modified buffer" highlight background colour does not stretch to the full width of section C, but that's a minor thing.