Environment tweaks for WSL
environment: Add $WSL when running on WSL, currently containing the
kernel revision.  The Windows console has an imperfect emulation of an
xterm termcap, and one of the quirks is that "less -X" will leave the
screen in reverse video when less exits.  Solve that for now by dropping
the -X from $LESS; hopefully the console's xterm emulation will improve
and/or they will customise the termcap to more closely match the
behaviour.