Commands to manipulate patches.
vim/autoload/diff.vim:
vim/startup/diff.vim: Commands to get a list of files from a git style diff, or diff -u patch. The use of a dict here to collect the list of files is an optimisation: the keys of a dict are a unique set, so maintaining the list of files as keys in a dict avoids the need to later uniq() the list.