Improve the git branch and commit info in statusline.
vim/after/plugin/statusline.vim: When navigating the git history with fugitive it is helpful to know whether the file you are looking at is the live working tree copy, or a previous commit. That is visible in the pathname, but it's usually so long it has been truncated. Instead let's add a shorter form of FugitiveStatusline() into statusline part B where I previously had the branch.