Checkintool and Gstatus/Gcommit helpers.
vim/startup/diff.vim: Add a CheckInTool command which will do the right
thing for the local repository: Gstatus for git and (if present) calling
CheckInTool() for others. Fugitive's gitcommit file type provides a
simple way to configure that buffer with spell enabled, and (if using
Gcommit --verbose or the cvc map from Gstatus) to read in the list of
files in the associated diff.