Command to open files from base64 encoded names
vim/startup/base64names.vim: Passing filenames which include shell
meta-characters is prone to problems even when using python's shlex as
run-in-terminal-from-finder does. Instead I'm going to change viterm to
base64 encode filenames and run vim -c 'EditBase64EncodedNames ...'.
This implements the new vim command. One thing to note is the buffer
number manipulations: the assumption is that if the current buffer is 1
then it's probably the empty file that vim opens with and so will be
reused for the first file opened, otherwise the first file will be one
greater than the current bufnr. A better option would be to check if the
current buffer contains a file or is empty.