Make bash handle output missing a trailing newline
functions.d/functions.bash: This is borrowed from zsh and the various
articles I've put in the comments. The idea is to output a marker and
then enough spaces to fill the line width and then a carriage return to
go back to column 0 before printing the shell prompt itself. If the
proceeding output was missing a newline, the marker will remain visible
but otherwise the marker will be overwritten by the prompt. Zsh actually
does the spaces+carriage return thing twice. This may not work with
shells with strange wrapping behaviour, and only works if output
directly to the terminal and not via $PS1 which bash may try and
interpret.

vim/vimrc: Update the EOL listchar to match the one I'm using in bash.
Though in bash it's an EOL missing marker.