environment: Fix degenerate macOS/iTerm paths
Somehow iTerm can end up with a degenerate path when started from Finder
by a workflow or Automator action. This is despite the script calling
osascript to start iTerm having the correct path in its environment.

So spot when the path might be degenerate (it doesn't start with
$HOME/*) and reset the path to the basic macOS path, then let it build
it up like normal.