Use $- for a robust check whether interactive
functions: It turns out that dash will always set $PS1 even in a
non-interactive shell. So the only safe way to test if the shell is
interactive is to look for 'i' in $-.

shrc: This whole file is essentially within "if interactive" so invert
the whole thing and bail out quickly if non-interactive, which
simplifies the code.