Add a customised tabline to vim
vim/plugin/statusline.vim: Expose the code to work out a sensible buffer
name as GetBufferName(bufnr_or_name) which can then be used from the
statusline code here and also from the new tabline. Also fix the
readonly marker ('RO') for vim :terminals which are always 'modified'
but also 'readonly'.

vim/plugin/tabline.vim: A new tabline implementation which allows for
tabs to be named with :TabName (or cleared with :TabName!) or falls back
on the active buffer name.