vim: Extend the simple BufExplorer alternative
vim/plugin/plugins.vim: Extend the simple BufExplorer alternative with a
:Bselect command that can make :bdelete simpler, and :Bcleanup a simple
cleanup of different types of buffers (invisible: not in a window,
nameless, and other: any other buffer but the current one).

Add Tim Pope's vim-eunuch which makes renaming the current buffer and
file so much easier.

Also a couple of mappings for Terminal-Job mode to make it easier to
switch to a different tab.