vim: make CheckInTool less surprising
vim/plugin/diff.vim: When called from vimr (via the |terminal-api|) the
message that the current directory is not a source repo could disappear
on a tab page that is immediately closed. Instead use echom to make it
stay around (and get added to :messages) and add a highlight. When
called from :CheckInTool without arguments explicitly use getcwd() as
the repo directory.