Vim fuzzy buffer lookup
vim/plugin/buffers.vim: Adding back some of the complexity now that I
understand how fzf --nth and --with-nth work. The format now includes
two tab separated fields, the first being a search key and the second
being the information to display to the user. The display format is now
similar to :ls.