Update brew scripts for clarity
mac/brew-list-orphaned-packages: I always was confused about what I
meant by orphaned, so switch the sense of the "in Brewfile" message and
make it so that by default it lists true orphans. sed -l ensures that
sed line buffers the output so it doesn't look like nothing's happening
for a long time.

mac/brew-list-reverse-dependencies: Restrict this to formulae since
that's what "brew uses" will work on.