Write a script to apply macOS defaults
mac/macos-defaults: A script to programmatically apply defaults to
macOS, e.g. for a new machine. By default this runs in read-only mode,
add --doit to run it for real and be prepared to be able to sudo.

mac/defaults/keyboard.language.json: A partial plist file in JSON format
to add Australian to the enabled keyboard layouts.

mac/defaults/keyboard.language.plist: The binary version of the same
partial plist file which is merged into
~/Library/Preferences/com.apple.HIToolbox.plist with PlistBuddy.