Fix vim error when opening a tar file
vim/plugin/buffers.vim: When opening a tar or tgz file it appears that
BufDelete can happen without either BufWinEnter or BufEnter preceding
it. This results in vim throwing an exception because t:buffer_list does
not exist, and the existing try...catch wasn't enough to deal with it.
So check if the variable exists before calling remove() but still retain
the try...catch to deal with any exceptions thrown by remove() itself.