brew-list-reverse-dependencies: Improve calculation
mac/brew-list-reverse-dependencies: 'brew deps --installed' does not
always seem to produce a full set of dependencies for the installed
formulae. Instead taking the list of installed formulae and then running
'brew deps --for-each <formulae>' works reliably. Then make use of the
list of installed formula to pre-populate the rdeps dict which makes the
code more explicit and maintainable.