Make .environment safe with bash -e
environment: In a script running with set -e, sourcing $HOME/.profile or
similar would fail because of the return 1's from add_to_path_var(). The
return value is only used by add_to_path_var_and_export() so make that
an internal detail and provide a wrapper that is safe for top level
callers.