mac/install: updates
mac/install -> mac/install.d: rename the directory of installation
scripts so that it's less likely to get in the way of tab completion.

mac/install.sh -> mac/install: rename the installation script and update
so it is clearer what is going on, and to make the handling of the
environment changing more robust.