Fix login shell detection under vscode
shrc: When a bash login shell is started from VSCode on Linux with shell
integration enabled, this was not detecting the login shell correctly.
The problem is that VSCode injects its own bash startup script with
--init-file and so this wasyn't seeing the shell as a login shell. The
solution is to look and see if .profile has been sourced prior to shrc.