vim: Fix shfmt format-on-save
vim/plugin/plugins.vim: Load vim-shfmt unconditionally so that it is
loaded before the related code runs.

vim/vimrc.sensible:
vim/after/plugin/shfmt.vim: Move the shfmt setup into an after/plugin
and implement a wrapper for format-on-save behaviour which tests whether
it is enabled at the point of execution.

vim/after/ftplugin/sh.vim: Remove vim-shfmt's autocommands if they
somehow got enabled. This shouldn't be needed, but it's idempotent so
better to be safe.