vim: Initial vim-lsp configuration
vim/plugin/plugins.vim: This is the bulk of the changes. Firstly this
disables older plugins which overlap with vim-lsp. Then it adds vim-lsp
and vim-lsp-settings, and sets up some configuration.

Configure pylsp to be more compatible with black formatting, and disable
some tools I don't need.

Set up clangd so that the choice of formatting style is based on the
repo being worked on, with the selection being cached.

vim/ftplugin/python.vim: Use black for formatting Python files. This
assumes that black has been installed in the environment. TODO: Add a
format-on-save toggle.

vim/vimrc.sensible -> vim/vimrc.2.vim: The "sensible" name was annoying
me so I renamed it to denote a stage 2.

vim/vimrc: Update to handle the new stage 2 file.

mac/install.d/ruby.sh: Install rubocop for Ruby linting.